refactor: Use netlink package to manage ifb interface
This change replaces the use of shell commands with the netlink package to manage the ifb interface in the linux/ip.go file. The modprobe and ip commands are replaced by the appropriate netlink functions to set up, tear down, and delete the ifb interface. This simplifies the code and improves the maintainability by relying on a native Go library instead of executing shell commands.

Signed-off-by: Yuriy Kis <yurakis9@gmail.com>